refactor(shared): extract relative-time helpers to fit source-structure ceilings
The previous commit pushed three allowlisted files past their frozen
line ceilings, failing check-source-structure. Rather than raise the
ceilings, the relative-time pair moves to its own utils/relativeTime.ts
module — time.ts keeps absolute formatting, parsing, and ranges.

HubWorkCard now imports through the declared ./utils subpath, so the
root barrel does not grow. Focus restoration moves into closeSpeedDial
in both FAB shells, which drops the duplicate call in the Escape path
and, because the dismiss backdrop now routes through it too, fixes a
stale focusedSpeedDialActionId on backdrop dismiss.
