+++ packages/shared/src/index.ts
@@ -1344,7 +1344,6 @@ export {
   getNetworkContracts,
   getOctantVaultRedeemCallShape,
   getOpenInBrowserUrl,
-  getRelativeTimeParts,
   getReviewedActionTranslation,
   getRoleColorClasses,
   getRoleLabel,
@@ -1385,7 +1384,6 @@ export {
   normalizeCampaignAddress,
   normalizeCampaignMetadataUrl,
   normalizeDecimalInput,
-  normalizeTimestamp,
   // Form normalizers
   normalizeTimeSpentMinutes,
   normalizeToFile,
